Improve input traffic efficiency
The rate at which events were fired caused too much of a load for the
system to handle, so I had to come up with a better solution than
sending the raw input information.

The way we sent input data now is as follows
1. When checking the event queue we only use the axis data of the last
item in the queue. This is because the entire arc of movement for each
input causes an information overflow and it doesn't provide anything
useful. The only thing that matters, for anything that cares, are the
current states of each axis

2. Buttons are the opposite. We need to track that a button was clicked
and then deselected. So we pull from the front of the event queue and
then remove any consecutive duplicate inputs for the next loop of
the schedule.