Smooth out mouse motion
The movement of the mouse needed to be a little smoother. Now, when
a mouse command is received, the command is processed on a background
thread to incrementally position the mouse based on the refresh rate
of the thread. The thread will kill itself off once the velocity of
the mouse is zero to save on processor energy.